move missile guidance to subaddon
