+++ addons/compat_cwr3_baf/compat_cwr3_baf_missileguidance/script_component.hpp
@@ -1,0 +1,4 @@
+#define SUBCOMPONENT missileguidance
+#define SUBCOMPONENT_BEAUTIFIED Missile Guidance
+
+#include "..\script_component.hpp"
